+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 0.3.0
+- Added `PageletRails::Controller` concern which allows to use pagelet functionality within existing rails controllers
+- Added `pagelet_method` method which enables pagelet functionality for a single method
+- `PageletRails::Concerns::Controller` renamed to `PageletRails::Component`
+
 ### 0.2.2
 - Fix links with data-remote=true wrongly selected ([#18](https://github.com/antulik/pagelet_rails/pull/18))